[ADDED] FileStore: option to limit use of File Descriptors
Each channel corresponds to a sub-directory and each
sub-directory contain several files. The more channels are
added to the system, the more FDs are needed.

This option is a soft limit instructing the store to
try to keep the number of opened files below this limit.
It is possible that during concurrent read/write across
multiple channels the number of FDs goes above the limit.

Note that this is for the file store's FDs, not the server
as a whole. For instance, if the server embeds the NATS
server, the number of FDs for the nats-streaming-server
process could be much higher (additional FD per Streaming
client connecting to the server). If that's an issue, one
should configure the nats-streaming-server so that it
connects to a remote NATS Server.